Silentbot-AI-main/Silentbot-AI-main/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
jinja2
rich
